Bug #20342004 SERVER_STARTUP_SHUTDOWN_TIME UNSTABLE ON PB2
This test measures time taken by the server to start and shut down. This timeis dependent on the operating system, hardware specs, load on the machine when test runs etc. It is not possible to have stable runs of this test in a PB2 setup. Remove test from MTR suite and replace with a test in performance suite.

Approved by Erlend Dahl <erlend.dahl@oracle.com>