Bug #23699294:ROWS EVENT ARE DECODED ONE BYTE AT A TIME
Analysis:
The constructor of Rows_events takes a buffer that contains the
row event and decodes it into fields. One of these fields is the
actual row data, which is being copied from the buffer one byte
at time, which is not required and is a slow way to copy data.

Fix:
Copy the complete data at once.