Bug #23522027 NDB : READ BACKUP NOT WORKING FOR ORDERED INDEXES
Ordered indexes inherit the ReadBackup property of the indexed
table, but this was not implemented, causing OrderedIndex scans
to continue to be routed to Primary fragment replicas.

DICT is modified to set the ReadBackup property on Ordered indexes
from the main table property, when distributing information to
TC and SPJ instances.

A new testcase, ndb_read_balance is added, which shows this problem,
and some others.