Bug #23633848 NDB : READBACKUP NOT WORKING WITH SPJ ORIGINATED REQUESTS
The SPJ block was continuing to read Primary fragment replicas, as shown
by the new ndb_read_local testcase added in thie commit.

This was due to a bug in the code checking data from on DIH on whether
a local backup fragment replica was available.

This bug is fixed.