WL#7069: Followup: Extend deserialization to handle both const and non-const cross reference pointers
Problem: The functions overloads lookup_opx_reference(Sdi_rcontext*, T**. uint) would not bind their second argument to &var where the type of var is const T*.

Solution: Change the function so that it returns T* and let the templated
calling function perform the assignment. Since the calling function is
templated, overloading must still be used to select the correct version of the
function. Since it is not possible to overload on return value, a dummy
const T* argument must be added to the signature. This works because const T*
will bind to both a T* and const T* (unlike T** or const T**).