Bug #23747899: @@BASEDIR SYSVAR VALUE NOT NORMALIZED IF SET THROUGH   THE COMMAND LINE/INI FILE
The system variable basedir is of type charptr and can take command line
arguments. This means that it operates on a "char *" global and every time
a command line argument is supplied a new string buffer is allocated and
assigned to this variable.
The global char * used for this variable is called base_dir_ptr.
Originally this char * is assigned to a global char array called base_dir.
And the rest of the server code uses the base_dir array directly.
But setting a new value on the command line set the base_dir_ptr to the
something different than the base_dir.
Then, in mysqld_get_one_option(), the contents of that new buffer is copied
to base_dir, but the base_dir_ptr is not reset to point back to base_dir.
Thus it does not reflect the subsequent processing done directly over base_dir
This processing includes (among other things) normalization of the path
separators used. As a result the system variable returns the data exactly as
set on the command line or through the INI file instead of the normalized
value that's stored in base_dir.

Note that no other code uses base_dir_ptr, thus the problem is constrained
to the value returned for @@basedir.

Fixed by reseting base_dir_ptr to point to base_dir in mysqld_get_one_option
right after copying the new value back. This makes it similar to the other
path containing system variables.

Test case added.