Bug#23060553 MTR SPAWNS ONLY ONE WORKER WHEN USING --PARALLEL=AUTO              ON WINDOWS
Issue:
------
MTR does not compute the number of parallel workers when the option
--parallel=auto is given on Windows and spawns only one worker by
default. An additional check throttles the number of parallel workers
to one in case of virtual machines on Windows.

Fix:
----
Number of parallel MTR workers is now set to the number of processors
using the environment variable 'NUMBER_OF_PROCESSORS' when --parallel
=auto is given on Windows. Also, a line which checks if a Windows machine
is a virtual one and sets the --parallel value to one has been removed
as Windows VM's are stable with huge amount of threads these days.

Reviewed-by:
Sayantan Dutta <sayantan.dutta@oracle.com>
RB #12523