Bug#23714929: MTR TREATING --BINLOG_FORMAT CASE SENSITIVE WHERE AS SERVER DOES NOT
Issue:
------
When tests which include have_binlog_format*.inc are run with a particular
binlog format provided in the command line, the test gets skipped if the
value specified is in upper case letters.

Fix:
----
MTR has a list in which the binlog formats are stored depending on the
have_binlog_format*.inc included in the test which it is running. When the binlog format
is passed through the command line, MTR compares it with the elements in the list, and
skips the test depending on which binlog format the test supports. Since this list
contains lower case letters, the comparison is made case insensitive, so that the test
does not get skipped when valid values of binlog format with upper case letters
are provided.

Reviewed-by: Srikanth B.R <srikanth.b.r@oracle.com>
RB: 13316