Bug#21807248	MY_PRINT_DEFAULTS , MYSQLD NOT FOUND ERRORS FOR COMMUNITY TO COMMERCIAL UPGRADE
Upgrading from community to commercial works by removing community, then installing commercial.
Removing community doesn't remove the service script, which will be run as part of commercial's
pre-installation script to ensure no server is running. This resulted in error messages in the log
since it was trying to use mysqld binaries that had been removed.

Added checks to the usage of mysqld and my_print_defaults in the systemv service script.