Bug#21286722 MISSING SANITY CHECKS FOR MALLOC() IN RESTORE.CPP, DBUTIL.CPP, SQLCLIENT.CPP
PART 2: DBUTIL.CPPRT 2: DBUTIL.CPP

In DBUTIL.CPP we create buffers using system function malloc but we do not
check for the failure. As this may lead to crash, have added validation
to check the pointer for NULL.

In DBUTIL.CPP we create buffers using system function malloc but we do not
check for the failure. As this may lead to crash, have added validation
to check the pointer for NULL.