Bug#22353594 Synchronize clusterj dbs hash map to avoid possible race condition
The dbs hash map maintains a collection of each DBImpl object (Ndb wrapper)
given to the application. This change synchronizes the map to avoid
possible simultaneous get and remove operations conflicting.
The result of a conflict may be to miss removing or adding
an object, which could make the number of db objects as seen
by clusterj to be different from the number seen by the application.