Bug#24300175 - IMPROPER CALLS TO TRX_SEARCH_LATCH_RELEASE_IF_RESERVED
After we have more than one btr_search_latch, we don't have to keep
this latch after we return from row_search_mvcc. So any other check
to release the search latch if reserved are unnecessary for now.
There are some unnecessary and contradictory code left. So we have to
remove them.

Reviewed-by: Jimmy Yang <jimmy.yang@oracle.com>
RB: 13400