Bug#24293015: HEAP-USE-AFTER-FREE IN MAIN.FLUSH_READ_LOCK
The problem was that SHOW PROCESSLIST could end up reading
freed memory when accessing THD::m_query_string belonging
to a connection that was in the process of deleting a
Prepared_statment (e.g. due to reprepare).

The reason was that THD::m_query_string could be allocated
on the wrong MEM_ROOT - not the statement MEM_ROOT, but the
MEM_ROOT belonging to the prepared statement. So that even
if SHOW PROCESSLIST locks the mutex protecting
THD::m_query_string, the prepared statement could make the
string invalid by deleting Prepared_statement::main_mem_root.

This patch fixes the problem by making sure that
THD::m_query_string is allocated on the correct memroot
before the prepared statement terminates.

The patch also includes some minor refactoring
related to 'expanded_query'.

Change-Id: I54119f17f1549d8856da5374f9d68cc1442d8eb9