Bug#24006602 	INNODB.SHOW_ENGINE_STATUS FAILS BECAUSE RW-S LOCK ROUND COUNT DOES NOT INCREASE
This test was introduced by rb#12420 and started failing frequently on PB2 because it was not correct.
Sometimes, when we can have lock already, we don't spin at all, hence the count does not increase. What we want to check in this test is whether the count decrease (it would take place if the counter overflow).

RB#13445
Reviewed by Kevin Lewis.