Bug#24330138 RPL.RPL_BINLOG_ERRORS FAILS SPORADICALLY ON WINDOWS
Problem and analysis:

The second part of the test case is doing a RESET MASTER to remove all binary
log files and generating a single new binary log file. Then, the test case is
simulating a binary log issue to make a FLUSH LOGS to not generate a new
binary log file. The test case asserts that there is only a single binary log
file on the data directory after the FLUSH LOGS.

The assertion is using the following MTR command:
--list_files $MYSQLD_DATADIR *master-bin.0*

On Windows, it is possible that the old binary log cannot be removed promptly
and the old file will be renamed to "master-bin.000001.<SOMETHING>.deleted". The
problem is that the "deleted" file name will also match the assertion above
described, making the test to fail.

Fix:

Used a pattern that will not display deleted binary log files on the test
case.