Addendum to bug #24311444: proper destruction of the suid view Security_context instances
The view Security_context object instances themselves are allocated
out of a mem root, but they allocate extra dynamic storage for
their members from the heap.
This extra dynamic storage needs to be released when the
Security_context instances are destroyed.
But since they are allocated from the mem root we can't just call
delete for them.
So we call the destructor explicitly to clean up any transient
heap bound memory that could have been allocated by the instances.