Bug #19474782: NDB_LOGEVENT_GET_NEXT DOES NOT RETURN ERROR IF MGM CONNECTION IS LOST
Problem:

When an mgmapi program that continuously monitors events is being run,
and if the management node is restarted, the events thereafter are not
reported to the mgmapi program. Timeouts are returned indefinitely
instead of an error being reported.

Cause:

This happens because the sockets of event listeners are not being
closed when the mgmd object is being deleted while restarting
mgmd.

Fix:

Added code in the destructor of Ndb_mgmd_event_service to close
event listener sockets.
Now, Read Error is received after mgmd restarts.