WL#9468: Step 2. Create a template for a stateless allocators which can be instantiated with callable class types that use different performance schema memory keys.
The std::basic_string template can only be instantiated with allocators that
are default-constructible (See
http://gcc.gnu.org/bugzilla/show_bug.cgi?id=56437 "basic_string assumes that
allocators are default-constructible"). So what is needed is a stateless
allocator which can use different PSI_KEYS. In order to be stateless the
PSI_KEYs have to be provided as template arguments. Unfortunately, this is not
possible, as PSI_KEY values are not compile time constants.

This changeset introduces a Stateless_allocator template which takes two type
arguments that are callable objects (functors) which will be used to allocate
and deallocate memory. This works because the callable object type is known at
compile time, but the value of the PSI key is not resolved until the allocation
actually happens. The drawback is that a separate Functor class has to be
created for each PSI key which will be used with the Stateless_allocator template.