Bug #24353408 NDB : NDBMTD ERROR FILE REPORTS WRONG FILENAME FOR THREAD 0 TRACE
Problem:
--------

In case of an ndbmtd crash, the error log output specifies the trace
file name which includes the thread number as a suffix to the name.
However, the trace file for thread 0 is actually named without a _tX
suffix whereas an incorrect _t0 suffix is added in the ndb error output.

Fix:
----

Ensure that the _t0 suffix is not added to the trace file name
in the ndb error output for thread 0 trace.