Bug #23748958 NDBMTD CRASH ON CREATE NODEGROUP WITH INVALID NODE ID (PTR TOO LARGE)
When DBDIH receives a request to create a nodegroup, it checks
the status of the node IDs passed in the CREATE NODEGROUP command.
Out-of-range node IDs result in a crash since there is no bounds
checking before node ID lookup. Added a check for out-of-range
node IDs for CREATE NODEGROUP in DBDIH. Also added a new error code
for invalid or undefined node IDs.