Bug#24365738: MYSQLTEST MAY FAIL TO COMPILE BECAUSE               MYSQLCLIENT_ERNAME.H NOT YET GENERATED
The problem was that building of mysqltest could fail
since it depended on GenClientError (and GenError) and these
targets might not have finished.

This patch fixes the problem by adding the missing
depedencies to mysqltest.

This was a regression caused by the patch for
Bug#21048973 SUPPORT "--ERROR CR_..." FOR COMMUNICATION RELATED ERRORS