Bug#22987899: ALTER TABLE MODIFY COLUMN DOES NOT SET               ALTER_CHANGE_COLUMN_DEFAULT FLAG
Change the parser to the ALTER_CHANGE_COLUMN_DEFAULT flag
is set not only for ALTER TABLE ... ALTER COLUMN ... SET DEFAULT
but also for ALTER TABLE ... CHANGE COLUMN and MODIFY COLUMN
if the DEFAULT clause is used.

This means that inplace ALTER TABLE is now possible for
column modifications on tables which have generated columns
depending on columns with DEFAULT values if the modified
columns are not involved in the generated expressions.
For example changing the NULL property of a separate column
can now be done inplace without table rebuild.