Bug#23742016: MTR --PARALLEL=AUTO IS ALWAYS 1 ON OSX
Issue:
------
MTR currently checks for CPU information using '/proc/cpuinfo' on
linux and 'kstat' on solaris. As no information can be obtained
from them on OS-X, a dummy CPU is added and the number of CPU's is
reported as 1 by default. Hence, using --parallel=auto leads to
MTR running tests on a single worker.

Fix:
----
A routine has been added to get CPU information on OS-X using the
'sysctl' command. The number of parallel MTR workers is set to the
number of processors in the machine.

Reviewed-by: Sayantan Dutta <sayantan.dutta@oracle.com>
RB: 13389