Bug#23028418 - RESTARTING DATA NODE CRASHES CLUSTER; NDBREQUIRE FAILS WHEN CREATING DICT TABLE
When a data node has insufficient redo during a System Restart(SR), it
does not participate in the SR until after the other nodes have
started.  Then it performs a Takeover of its fragments from the
started nodes in its nodegroup.  During this takeover the node still
considers that it is involved in a System Restart (SR), but it is more
akin to a Node Restart (NR), as the cluster is running and accepting
user transactions, DDL operations etc...

In SR scenarios, table creation handling in DIH has a different
behaviour than normal, as table 'creation' is related to reloading
table definition data from disk on the Master node.  During NR /
Takeover, table creation can occur on non-master nodes due to user
activity.  However, the mechanism in DIH to determine how to handle
takeover assumes that if the restart type is 'System Restart', then
any table creation must be SR related.

The solution is to add an extra check to this SR special case path to
look at whether the running node is the Master node and use that to
determine the difference between SR and (SR+Takeover).