Patch for bug#20402364: NDB_ADDNODE_WITHBINLOG FAILS DUE TO EVENT API STALL
Fixes various issues related to 'out of order' (OOO) completed event buckets:

1) Change nodegroup w/ SUB_CNT (remove nodes):
   Newer epochs which becamme complete after removing
   the nodes were not detected as out-of-order completed
   (as reported in the bug reported).

2) Change nodegroup w/ ADD_CNT (add nodes):
   Newer epochs need to wait for sub_gcp_complete from
   newly added nodes are falsely considered as complete

3) Pure OOO of OOO where 'm_latest_complete_GCI' is overwritten
   by a lower epoch completing later: While waiting for epoch
   'n' to complete, 'n+2' complete OOO, then 'n+1' causing the
   OOO' completed 'n+2' to be overwritten.

Whenever nodes are added or removed, the event-API is
notified by a SUB_GCP_COMPLETE_REP with either the
SUB or ADD flag set, and number of nodes to ADD/SUB.
This allows the Event-API to correctly maintain the number
of SUB_GCP_COMPLETE_REPs needed to be received in order to
complete a bucket.

When receiving a 'nodegroup ADD/SUB' the bucket fixup
is handled by NdbEventBuffer::handle_change_nodegroup().
In addition to handle the bucket for the epoch associated
with the SUB/ADD, it also has to adjust the 'm_gcp_complete_rep_count'
of the buckets assiocated with newer epochs arrived 'out of order'.

As mentioned in 1), that could possible complete
any such 'out of order' buckets. However, the code for
annotating the buckets as 'out of order' completions
was missing! Thus event reception effectively stalled.

This patch adds detection and handling of out of order
completion as part of ::handle_change_nodegroup().
(issue 1) & 2))

It also adds protection against 'out of order' completion
arriving 'out of order' in itself. For example if epochs
completes in the order n+2, n+1, n, the second out of order epoch
will set 'm_latest_complete_GCI = n+1', thus loosing the
completion of 'n+2' causing a later stall in the event-API
(issue 3) )

Also fixes some const correctness in related code.