Bug#24399819 SQL/RPL_WRITE_SET_HANDLER.CC INCLUDES PRIVATE LZ4 HEADER FILE
Problem: The fix for bug #23607230 used xxhash symbols from liblz4,
but even though the xxhash symbols are exported by liblz4, the header
file is not part of the API, so compilation fails when building with
WITH_LZ4=system.

Fix: Build xxhash separately from liblz4 so that it's available both
when using system and bundled lz4 libraries.