mysql-trunk-meta-sync:
Fix mysqldump_max_recordsize.

A COMPRESSED table could not be created when page size > 16K, InnoDB will
change the page format to DYNAMIC silencely.

Before writing metadata to global DD, server would still remember the
COMPRESSED. But now, DYNAMIC would be written back, so the table is in
DYNAMIC format, which can be accepted by a server with page size of 64K.
That's why it fails. Since we indeed change the format to DYNAMIC,
the test case should be adjusted.