Bug#22818781: EXPLAIN_JSON_VALIDATE.TEST FAILS ON WINDOWS PRE-PUSH TESTING
Issue:
------
The explain_json_validate test calls a python script which accepts arguments.
If the arguments are passed with wildcard characters, Windows cannot parse
them and hence the test fails.

Fix:
----
The fix involved adding an additional check if the test was running on Windows
so that the arguments can be parsed correctly.
Also, Python is not installed on Windows by default, and therefore the
script is now ported to perl. Before the test used to be skipped if the machine
did not have either Python or JSON support, but now it fails if it does not
have JSON or perl support.
Since the python script has been removed, all the tests which were using that
script are now using the perl script.

Reviewed-by: Horst Hunger <horst.hunger@oracle.com>
Reviewed-by: Guilhem Bichot <guilhem.bichot@oracle.com>
RB: 12537